set bind and queue in inherited callback
This basically copies the current state of the class level instance
variables to the subclass.

The inherited callback seems to work slightly different in ree and ruby 2+
but the current solution works both cases. If I understood it correctly
in ree and I assume 1.8.7 in general the inherited call for dynamically
created classes is done at the end of the class creation meaning the bindings
defined in the subclass are already present. In ruby 2+ for all classes and in
ree with static classes the callback is done at the beginning or at least
before the `bind` and `queue` calls from the subclass are happening.